# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DS ESTIMATION BACKEND DOCUMENTATION",
    "description": "Our story started in 2015 when there was a need of Construction database for helping bidders to bid accurately with reliable information's and estimates in the Rwanda construction industry. Continuous determination, hard work and passion have been constant themes throughout our history and present. As Rwanda Construction industry becomes increasingly important on the global stage, it's our believe that, with our own innovative mindset, we can make the construction industry a better one for national development. So we challenge ourselves to evolve, innovate and experiment to make work easier and faster for everyone as the leading local software brand in the Rwanda construction industry. We are currently the quantities and estimations software dealers which are made in Africa its self initially from Ghana and same technology has been customized to fit Rwanda construction industry and we are super proud to start the race with national development in minds. Today we are privileged to connect with thousands of Engineers and other construction related professionals who do daily estimates or quantities in the construction industry.",
    "version": "1.0.0",
    "contact": {
      "name": "DS ESTIMATION",
      "url": "http://dsestimation.com",
      "email": "dsestimation@gmail.com"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "AuthToken": {
      "type": "apiKey",
      "name": "auth-token",
      "in": "header",
      "description": "The token for authentication"
    }
  },
  "security": [
    {
      "AuthToken": []
    }
  ],
  "paths": {
...
    "/owner/material/create": {
      "post": {
        "tags": ["Default Material Routes"],
        "summary": "Create a material",
        "description": "Create a material",
        "operationId": "createMaterial",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Material object that needs to be created material",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Material"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Material created successfully"
          },
          "400": {
            "description": "Some material field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/owner/material/{id}/update": {
      "patch": {
        "tags": ["Default Material Routes"],
        "summary": "Update a material",
        "description": "Update a material",
        "operationId": "updateMaterial",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the material to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Material object that needs to be updated default material, you can update one or more fields",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Material"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Material updated successfully"
          },
          "400": {
            "description": "Some material field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Material not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/owner/material/{id}/delete": {
      "delete": {
        "tags": ["Default Material Routes"],
        "summary": "Delete a material",
        "description": "Delete a material",
        "operationId": "deleteMaterial",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the material to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Material deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Material not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/user/material/all": {
      "get": {
        "tags": ["Client Material Routes"],
        "summary": "Get all materials",
        "description": "Get all materials",
        "operationId": "getAllMaterials",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Materials Retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/user/material/create": {
      "post": {
        "tags": ["Client Material Routes"],
        "summary": "Create a material",
        "description": "Create a material",
        "operationId": "createMaterial",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Material object that needs to be created material",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Material"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Material created successfully"
          },
          "400": {
            "description": "Some material field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/material/{id}/update": {
      "patch": {
        "tags": ["Client Material Routes"],
        "summary": "Update a material",
        "description": "Update a material",
        "operationId": "updateMaterial",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the material to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Material object that needs to be updated Client material, you can update one or more fields",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Material"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Material updated successfully"
          },
          "400": {
            "description": "Some material field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Material not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/material/{id}/delete": {
      "delete": {
        "tags": ["Client Material Routes"],
        "summary": "Delete a material",
        "description": "Delete a material",
        "operationId": "deleteMaterial",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the material to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Material deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Material not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/global/material/all": {
      "get": {
        "tags": ["Global Routes"],
        "summary": "Get all materials from our database",
        "description": "Get all materials from our database",
        "operationId": "getAllMaterials",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Materials Retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Material": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "quantity": {
          "type": "number"
        },
        "unit": {
          "type": "string"
        },
        "caveragePerUnit": {
          "type": "number"
        },
        "price": {
          "type": "number"
        },
        "supplierId": {
          "type": "number"
        }
      }
    },
...
  }
}